Add Trino engine test config to SQL Engine Tests action
This adds a Trino test job to the MetricFlow SQL Engine Tests
action.

Trino can run its tests in a service container without need for
an external environment, however, the service container execution
for these tests is materially slower in CI runs than Postgres.

Rather than delay PR-blocking unit test completion we put the Trino
test suite execution in the label-initiated SQL Engine test action.
This allows us to move Trino to cloud-hosted execution, if we should
care to do so, but for the time being we will stick with the in-memory
service container.