Remove MarkupSafe dependency
We do not use MarkupSafe directly. Historically, we've needed this
because sometime, somewhere, something was causing runtime problems
with the way Pandas was importing and loading MarkupSafe.

Whatever that was appears to have been fixed - clean installs still
load MarkupSafe (since it's a transitive dependency) and everything
appears to be working as normal.