Update dbt-metricflow version to 0.5.0
With the release of MetricFlow 0.204.0 we can now release
dbt-metricflow 0.5.0 to complement it.

Note MetricFlow nominally depends on dbt-core 1.7.x, but installs
with that specific version will not be able to use conversion
metrics. That's a boundary case we should probably fix with a patch
release for MetricFlow.

In the meantime, we  ensure the installation helper dbt-metricflow
will force an update to the proper minimum version of dbt-core.