Use log level info when populating persistent source schema
Databricks and BigQuery are super slow on data load, which leaves
us wondering if they're doing anything. Without a log-cli-level set
the action logs just sit for a good 3 minutes. This adds log-cli-level
info to the command in order to provide some more detail on what the
actions are doing, which will also be helpful if something should fail.