Remove custom changeFormat from Docs changelog entries
The Docs entry appears to be a template meant to be used for
changelog entries attached to PRs where corresponding docs
updates are being cross-linked, but this is not the changelog
process we currently use within MetricFlow. Consequently, someone
adding a `Docs` changelog entry for updates to, e.g., our
CONTRIBUTING.md file would end up pointing the changelog entry
at an incorrect or outright missing issue entry in the docs repo.

If we update our changelog process we'd be better served with a
more obvious separation between repo-local documentation updates
and dbt docs site updates, so for now we simply remove the
Docs entry customization.