Suppress changelog links for missing Issues
The `Issue` entry in our changelogs is required to be non-empty,
and if it is we format the value into a GitHub link in the final
changelog. This leads to broken links pointing at things like
issue `None` or `0`.

This change updates the changelog formatting to skip over those
entries when forming issue links, and skip formatting in issue
lists if no issues are included. We still include the contributor
and change body, so all due credit remains, just minus the need
for validating common circumventions to our request for issue
numbers.

Running `changie batch 0.205.0  -d` on the underlying trunk commit
and comparing it here illustrates the new behavior with the
docs entry - and the associated author change list - no longer
including bad links.