Improve error message for metrics with no parents
Due to a gap in our upstream model validation it is possible for
a user to define a derived metric with no inputs. This causes a
fall-through error to trigger of the form:

`unbound method set.intersection needs an argument`

The root cause of this is our use of the `*` operator for expanding
an iterable into a set of values. If the iterable is empty, Python
will simply pass nothing - not None, but nothing at all. This means
any call like `set.intersection(*input)` will raise this strange error
if `input` is empty.

This change does two things to resolve this issue:

1. It updates the set.intersection(*input) call to a more robust
approach which will return an empty set if the input is empty
2. It updates the invocation to check for empty input and include
a more refined error issue indicating the root cause.

The reason we do it this way is the offending call is inside of
a more generic helper, so we allow the helper to accept empty
input and return correspondingly empty output, but at the callsite
where we know the input should not be empty and that empty input
indicates a configuration issue we raise the appropriate error.