Remove unsupported legacy MetricFlow Client API
In the original version of MetricFlow we supported a Python
client for programmatically issuing queries. This was a thin
wrapper over our internal APIs for running queries against
SQL engines.

Since the acquisition of Transform by dbt Labs the direction of
MetricFlow has shifted heavily (and appropriately) in the direction
of being a support library for building and rendering metric queries
against target query dialects (currently all SQL).

This legacy public API has not been maintained at all - it uses legacy
string-typed inputs, it does not support saved queries, and it allows
for a highly limited scope set of operations.

Furthermore, it is in a package directory path that indicates we have
a public API. Since we do not, in fact, have a supported public API
just yet, it behooves us to remove this in order to unblock some
efforts to simplify our internals and move us closer to the point
where we can build out a publicly supported API, whenever that time
may be.