Remove unused ci configuration file
The validation duration file was a temporary benchmarking action
that is so old it still references poetry commands, which have not
worked inside the MetricFlow repo for nearly a year now.

This was uncovered in a rename operation associated with forking
the test package out to repo root, so we just tidy it up here.