Exclude unused package components from metricflow build
The MetricFlow build was taking up 1.5-3 MB on disk, a large
part of which was the test package. Confusingly, the MetricFlow
build also included the contents of dbt-metricflow, which is
meant to be a separate distribution wrapper for MetricFlow itself,
so that's odd.

This commit cleans up the exclude list to be more appropriate to
MetricFlow's repository structure, and finishes the job of reducing
the size of the MetricFlow distribution packages by a factor of 4.

Most of the work was handled incidentally by moving the tests under
the repository root tests directory, which happened to be in the
exclude list already.