Enforce consistency across lint environments
Some lint errors cropped up in CI which could not be reproduced
locally, perhaps due to platform-related differences in versioning
or parsing for the associated tools.

This commit seeks to resolve those discrepancies by doing the following:

1. Removing the file path restrictions in the .pre-commit-config.yaml
2. Marking dbt_metricflow as a known first party package for isort

All other changes are the result of running the linter with autofix
enabled and these new settings in place.