Update dbt-semantic-interfaces to 0.5.1
In order to proceed with development on using Metric inputs in
filter expressions, we need to update our dbt-semantic-interfaces
dependency to 0.5.1 or later.

This change carries some added complexity along with it, since our
dev dependencies specify dbt-core and dbt-<adapter> package
dependencies bound to dbt-core 1.7 version range. This creates
an unresolvable dependency version conflict, since dbt-core depends
on dbt-semantic-interfaces 0.4.x.

For the time being we can get away with relying on the pre-installed
dbt core + adapter packages pulled in by dbt-metricflow, and
simply allow metricflow's specification for dbt-semantic-interfaces
to override the installed package.

Most of the changes in this commit are a result of needing to manage
the dbt-semantic-interfaces version discrepancies by relying solely
on the dbt-metricflow pre-install for dbt core + adapter dependencies.

The one additional tweak is the MetricFlow minor version, which was
improperly set to the currently available release of 0.205.0. Since
this is the top of a set of serious breaking changes we need to ensure
we release as a full minor, so the next version will be 0.206.0.