Remove pointless loop over DataflowPlan optimizer sets
The DataflowPlan optimizers are, at least in theory, applied as part
of a loop of optimizer sets to try, with each iteration of the loop
theoretically applying a different set of optimizers, and concluding
with an empty set. The idea behind this was quite clearly to ensure
that the optimizers don't cause valid queries to fail due to some
incorrect optimization path. Why we deemed this necessary is completely
unclear, as no documentation has been provided to explain this bit
of complexity.

What was happening in practice was that on every iteration we were
applying the same hard-coded set consisting of the SourceScanOptimizer,
and the loop itself would incorrectly report error logs suggesting
that other things were being attempted.

Since this logic has never been applied correctly, the optimizer
parameterization is buried so deeply into an implementation
method that there's no way to access it via any reasonable mechanism,
and the built-in loop iteration makes everything here look more
complex than it really is, we remove the loop here.

Note this was added fairly recently, but since the reasoning for
the change was never provided it appears that what amounts to a
revert of the salient commit from PR #976 should be fine.