Remove unused output_table parameter from MetricFlowQueryRequest
The output_table parameter in MetricFlowQueryRequest is a vestigial
holdover from back when MetricFlow was attempting to be a semantic
layer in a box. As that dream has foundered on the rocky shores of
Python dependency management, and MetricFlow itself has emerged from
the wreckage as a query building and rendering library, it makes
sense to remove some of these legacy execution layer elements.

This simply removes the parameter and the tests built around
exercising it. The parameter itself is not used in either
MetricFlow itself or in our cloud execution environment.