Consolidate linkable element from metrics accessors
The MetricLookup class has two separate methods for getting
linkable elements for a set of metrics. One returns an ordered
sequence of LinkableSpec instances, while the other returns the more
detailed LinkableElementSet container class.

The internal logic of both methods is identical apart from the
mechanical conversion to an ordered sequence of spec instances, so
we consolidate them into a single method and do a bit of renaming
and updating associated with the change. An astute reader may notice
that the more restrictive method is not used anywhere outside of tests.