Make MetricLookup linkable spec accessors return LinkableElementSet
When we do group by item resolution we rely on the MetricLookup class to
extract the set of linkable elements available for a given measure or
metric-free query. This gives us the full set of LinkableInstanceSpecs
that we can then use to build our query plan and convert to SQL for
rendering.

The current accessors all return the bare minimum set of information we
require - the specs themselves - and those get filtered and merged
inside of the group by item resolution process.

Predicate pushdown changes our requirements, because we suddenly need to
know, for any where filter, where all of the questionably named
"group by items" associated with the filter expression originated from.
This information is actually available in the LinkableElementSet the
MetricLookup retrieves, but it's not proapagated through the relevant
API methods.

This updates the methods in MetricLookup and pushes the transformation
logic from LinkableElementSet -> Tuple[LinkableInstanceSpec] out to
the callsites. Some minor method renaming and documentation updates
are included for clarity around the function of these methods.