Add tests for MetricLookup linkable element extractors
These methods in the MetricLookup class have behavior that
is somewhat difficult to reason about, and having a test that
both illustrates the output expectations for a given call and
provides some assurance that these accessors retain consistent
behavior is worthwhile.

This commit adds the missing test with a small filter. If this
becomes noisy we should move the semantic model container tests
off of the simple model and onto something less subject to change
(and also something with fewer join links).

The test for getting linkable elements for measure is also moved
under the metric accessor test suite, because the current API
sequencing suggests that the MetricLookup class is what we wish
to rely on, while the ValidLinkableSpecResolver implementation is
more fungible. This way, we ensure that the method we actually
use is covered without duplicating the pass through test in two
locations.