Refactor logic for finding available LinkableMetrics
Update logic to ensure we cover all allowed join paths. I separated this logic from the other linkable element logic, since there are more join paths allowed for LinkableMetrics than other elements. Here we can allow joining across any semantic models that share an entity, since fan-out will be prevented by metric aggregation on the right side of the join.
This also changes how we classify LinkableMetrics. Previously, we considered them local if they joined via an entity in the measure's semantic model, and single-hop if they required an additional join. Now, they will be classified as single-hop in the first case, and multi-hop in the second. This seemed more accurate, since it actually counts the number of joins required to access the metric. Plus, now we have the logic in place for more hops if we ever increase the max number of hops allowed.