Split LinkableElementSet into its own module
This is now being returned through methods in the MetricLookup,
which means it can be used independently of the LinkableSpecResolver.

As such it merits its own module. This will also be useful for the
short term need to access this class in the predicate pushdown
implementation, and longer term might make it easier to consolidate
this class with the LinkableSpecSet.