Add LinkableElementType annotation to ElementPathKey
Currently the ElementPathKey is agnostic to the type of element
it represents. In some cases this will be fairly obvious - if the
time granularity is set it is a time dimension. In others we can
determine this by inspection - if the name and links resolve to
a distinct dimension in a semantic model then it must be a dimension,
and similarly for an entity.

However, there is the possibility of some ambiguity in rare cases where
an entity and dimension might end up with the same identifier/link
combinations. This doesn't matter in the current usage of the ElementPathKey,
which is always bound to an element subtype, but when we add support
for predicate pushdown we'll need to be able to use this key to fetch
the specific set of LinkableElement classes, and this means we must
be able to easily get a determinstic type.

This commit adds the elmeent type annotation we require, and does some
additional related work to ensure that time dimensions and dimensions
are always classified correctly when the LinkableDimension instance
is created.