Remove granularity-free ElementPathKey instances
The ElementPathKey object is primarily used for accessing the set
of linkable elements associated with a given semantic model element
and the linkable path used to get to it from a query or metric.

The accessors implicitly require that granularity be set on all
ElementPathKey instances associated with Time Dimension elements.
Passing in an ElementPathKey for a time dimension without a granularity
will result in it being mistaken for a categorical dimension, causing
the path key lookups to behave strangely.

Attempting to make this implicit requirement explicitly enforced
uncovered an area of the codebase using ElementPathKey as a deduplication
mechanism for TimeDimensions. This convenience usage is harmless in and
of itself, but if those ELementPathKey instances were ever used for
access weird things would happen, so we remove them in preparation for
adding the relevant enforcement.