Make linkable elements available in FilterSpecResolutionLookup
We need to access LinkableElements - or, more specifically, their
source SemanticManifestReferences - in the WhereSpecFactory in order
to evaluate fitness for predicate pushdown.

This makes these elements available to the FilterSpecResolutionLookup,
which we use to get the resolved filter elements when building the
WhereFilterSpec. Subsequent changes will plumb this through to the
WhereFilterSpec for predicate pushdown evaluation.