Map joinable metrics to entities instead of semantic models (#1160)
<!---
Include the number of the issue addressed by this PR above if
applicable.
  PRs for code changes without an associated issue *will not be merged*.
  See CONTRIBUTING.md for more information.
-->


### Description
This fixes a bug in the logic deciding which `LinkableMetrics` show up
as options for a given query. Previously, we mapped semantic models to
the metrics they could join to, but that wasn't specific enough.
Sometimes there are entities in a semantic model that can join to a
metric, while there are others that can't. Here we update the logic to
map `LinkableMetrics` to entities instead.
Also changes the way we categorize `LinkableMetrics` in terms of
`LinkableElementProperties`. Previously, we would assign properties
based on joins like this:
`1 join = (METRIC), 2 joins = (METRIC, JOINED), 3+ joins = (METRIC,
JOINED, MULTI_HOP)`
Here we update the logic to assign properties like this:
`1 join = (METRIC, JOINED), 2+ joins = (METRIC, JOINED, MULTI_HOP)`
We've discussed this already as a team, and it just seems to make more
sense since all `LinkableMetrics` require a join.

Note that at the top of this stack will be a test that loops through all
allowed linkable metrics and runs them through the query parser +
dataflow plan builder to ensure we ACTUALLY allow all the options we say
we do.
<!---
Provide context for the Pull Request here, including more details on
what
  is changing and why. Add any references and info to help reviewers
understand your changes, such as any tradeoffs you considered, and the
local
  test process you followed.
-->

<!--- 
  Before requesting review, please make sure you have:
1. read [the contributing
guide](https://github.com/dbt-labs/metricflow/blob/main/CONTRIBUTING.md),
2. signed the
[CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
3. run `changie new` to [create a changelog
entry](https://github.com/dbt-labs/metricflow/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)
-->