Push resolved linkable elements through to WhereFilterSpec
The resolved LinkableElements gathered during the filter spec
resolution phase of query processing need to be available to
each WhereFilterSpec so that we can determine whether or not
a given WhereFilterSpec is eligible for predicate pushdown.

This change simply adds the necessary property and wires these
elements through to where we will need to access them in the
DataflowPlanBuilder. This change proved a bit troublesome due
to the SerializableDataclass construct, which has some particular
requirements which required some slightly odd inheritance tagging.