Use query parser to build group by metric source node (#1164)
This is needed to resolve the appropriate join path for multi-hop joins.
Currently users won't be able to pass a join path for those, so we need
to build it ourselves.
Moved source node-building logic to query parser class to avoid circular
import errors. This change also required adding the query parser as a
dependency for the dataflow plan builder.
This doesn't change the output of any existing tests, but I have some
relevant tests coming in future PRs.