Add support for set type collections in the pretty-printer
Our snapshot files rely on consistent order for all rendered
collections, but we were assuming that all input objects would
be strictly ordered. Unfortunately, the fallback method of using
Python's pretty-printer does not allow for ordered sets, so
any inputs involving set or frozenset types will be fundamentally
unstable.

This expands our list-like formatted print handling to cover set
types, indicating the distinction between set and frozenset and
printing all values in a stable sort order. This does assume all
set values are sortable, which should be the case.

As you can see from the snapshot updates, this is a pre-existing
problem in the codebase. We've gotten away with it because all of
the sets in our snapshots have, thus far, all been of length 1. This
will change in coming updates.