Always use spec's qualified name as column name
Previously we were duplicating this logic, which made it easy to overlook updating the column name to use the qualified name. I updated this for all specs so that it will update automatically in the future. You can see in the snapshots that this changed nothing for most specs, except group by metrics (as expected).