Handle both join paths in `GroupByMetricPattern` (#1166)
The main purpose of this PR is to ensure we parse both join paths from
the `MetricCallParameterSet` and handle them in the
`GroupByMetricPattern`. This piece is a bit hacky because we currently
don't support a couple things that we eventually will: 1) multiple
dimensions & entities in the `group_by` param for metric filters, and 2)
an `entity_links` param on metric filters. That will get cleaned up in
the next iterations of this feature.
The second commit here updates the metric subquery entity path (on
various classes) to include the last entity link, where previously it
didn't. This is due to a bug I found in the case when you have no entity
links left for the outer query - which happens when you have already
completed all joins needed to get to the metric subquery.