Store entity links on `MetricSubqueryJoinPathElement` (#1167)
Store entity links on `MetricSubqueryJoinPathElement`.
Originally I thought this wasn't necessary since the entity links should
be a duplicate of the entities in the join path. Later I realized that
there is one scenario where they don't match the join path. We allow
querying local entities both alone and prefixed with any of the other
entities in the same semantic model. If you query a local entity with
another entity prefix, the entity links won't match the join path. To
account for that, we need to store both entity links and join path on
`MetricSubqueryJoinPathElement`.