Remove metrics that must be queried with `metric_time` from linkable metric options (#1168)
Remove metrics that must be queried with `metric_time` from linkable
metric options. That means cumulative metrics with a window or grain to
date, as well as any time offset metrics. This restriction can be
removed when we add support for using multiple entities/dimensions in
the group by for metric filters.