Update DFP join logic to account for group by metric source nodes (#1169)
There was a lot of logic in the dataflow plan builder that assumed all
source nodes and joinable nodes would be `ReadSqlSourceNode` or
`MetricTimeDimensionTransformNode` (as was previously true). This is no
longer true since adding `LinkableMetrics`, for which you might have a
`ComputeMetricsNode` as a source node. This comes with a lot of changes,
like the fact that the entity links might look different for this source
node than for others. Here we update the DFP building logic to account
for those nodes.