Don't allow LinkableMetrics for case when entity reference is in entity links (#1200)
### Description
Some users have run into a scenario where the entity reference can be
found in the entity links for some `LinkableEntities`. I haven't yet
been able to reproduce that error with our test configs. Releasing
metrics as dimensions is blocked on this edge case, so I'm skipping that
scenario here to ensure that customers won't run into it until when
using `LinkableMetrics` until we can debug further. This will allow us
to release the feature and handle this case in a follow up version.