Dedupe filter suggestions
Currently, metric filters have a parameter that's not captured in their input string (the outer query entity links). This means that you might end up with multiple input strings that look the same, so we need to dedupe those. These duplicates will look different once we add a new parameter to metric filters that allows users to specify the outer entity links. But there should be no harm in using a set here, either way.